seed particle bugs
fixed bug in pp_intro that was causing shut-down when seed particle information was not provided at the simulation start, but was provided for later times. Also fixed bug in cham_up for instantaneous injection of seed particles, because the number of particle size bins (num_asb) was not defined. Also updated the code for finding chemical scheme files and xml files when the user provides only the file name - the code will look for those file names in the same folder as the model variables file now. Updated the fig01_mod_var.txt file so that chemical scheme files and xml files can be found automatically.