removed the forced vapour pressure
removed the forced vapour pressure setting of HNO3 and ammonium nitrate from prop_calc and volat_calc. Instead, if users wish to, they may set the vapour pressure of these components in the model variables file. Note that this change means that PyCHAM should now be comparable to older versions of version 4, and the version 4.7 is incomparable.