added model variable inorg_part_flag
patch to allow user to set inorg_part_flag to 1 and therefore use a field-observation based estimation for the vapour pressure (convolving any activity coefficients different to 1) of inorganics to simulate gas-particle partitioning of inorganics, but this should only be used if the chemical scheme allows for HNO3 to transform to NH4NO3 (ammonium nitrate) in the particle phase. The default value for inorg_part_flag is 0 which means that vapour pressures of inorganics are set by the default vapour pressure estimation method.