Minor upgrade for saving and error reporting
PyCHAM now creates a directory to save results to before the ode solver is called (isnide user_input.py), which means that it is much more difficult for PyCHAM simulations running at the same time to be running the same simulation, since a simulation can't start when the results folder already exists. In addition, if the ode_solv.ode_solv function returns an error, e.g. 'problem is singular', rather than crashing PyCHAM, an error message is returned to the PyCHAM GUI.