included a simpler way of calling PyCHAM
for an application programming interface with less time needed to set up the graphical user interface and quicker transitions to the next simulation in a batch, I created the PyCHAM_lite folder. In the future I will need to merge this folder into the standard PyCHAM folder, but with the option available inside the PyCHAM folder to run PyCHAM as it is run inside PyCHAM_lite.